Refactored animation and ball guy modeling to blender model
I decided it's time to move the ball guy model and animation to a place
where it should be - which is .blend object rather than java code.
(composing a character from manually created spheres was fun at the
beginning, but with adding animations it just became bothersome).

Also started using blender2ogre exporter which magically fixed couple bugs
with models' lighting and deforming.