changed status display implementation
Previously I struggled with the way status bar was implemented, mainly because it was set up with static layout and every selected item had to conform to it and I had no idea how to fix it.

Now every Selectable is responsible for the way its' status is layed-out by the way of simply providing gui Node that is attached in proper place (when unit is selected) and updated (when needed).

This leads to lot more flexibility in way of how each status is displayed.