starting to get rid of spring context
it's turning out to be quite a disaster so I'm starting a painful process of reverting it.

First I created a GameState which is going to contain game-specific thingies, added it to context and switched usages of those thingies to use them via GameState.

Next, after all the usages will be removed, I'm going to switched beans using it to POJOs.